Moved add to frontmatter functions back to utils
