+++ format-importer.ts
@@ -170,38 +170,5 @@ export abstract class FormatImporter {
 		await this.vault.append(fileRef, '', writeOptions);
 	}
 
-	/**
-	 * Adds a single tag to the tag property in frontmatter and santises it.
-	 * Must pass in app.fileManager.
-	 */
-	addTagToFrontmatter(tag: string, fileRef: TFile, frontmatter: any) {
-		const sanitizedTag = sanitizeTag(tag);
-		if(!frontmatter['tags']) {
-			frontmatter['tags'] = [sanitizedTag];
-		} else {
-			if (!Array.isArray(frontmatter['tags'])) {
-				frontmatter['tags'] = frontmatter['tags'].split(' ');
-			}
-			frontmatter['tags'].push(sanitizedTag);
-		}
-	}
-
-	/**
-	 * Adds an alias to the note's frontmatter.
-	 * Only linebreak sanitization is performed in this function.
-	 * Must pass in app.fileManager.
-	*/
-	async addAliasToFrontmatter(alias: string, fileRef: TFile, frontmatter: any) {
-		const sanitizedAlias = alias.split('\n').join(', ');
-			if(!frontmatter['aliases']) {
-				frontmatter['aliases'] = [sanitizedAlias];
-			} else {
-				if (!Array.isArray(frontmatter['aliases'])) {
-					frontmatter['aliases'] = frontmatter['aliases'].split(' ');
-				}
-				frontmatter['aliases'].push(sanitizedAlias);
-			}
-	}
-
 }
